+++ src/pipecat/processors/frameworks/rtvi/__init__.py
@@ -1,0 +1,73 @@
+#
+# Copyright (c) 2024-2026, Daily
+#
+# SPDX-License-Identifier: BSD 2-Clause License
+#
+
+"""RTVI (Real-Time Voice Interface) protocol implementation for Pipecat."""
+
+from pipecat.processors.frameworks.rtvi.frames import (
+    RTVIActionFrame,
+    RTVIClientMessageFrame,
+    RTVIServerMessageFrame,
+    RTVIServerResponseFrame,
+)
+from pipecat.processors.frameworks.rtvi.models_v0 import (
+    ActionResult,
+    RTVIAction,
+    RTVIActionArgument,
+    RTVIActionArgumentData,
+    RTVIActionResponse,
+    RTVIActionResponseData,
+    RTVIActionRun,
+    RTVIActionRunArgument,
+    RTVIBotReadyDataDeprecated,
+    RTVIConfig,
+    RTVIConfigResponse,
+    RTVIDescribeActions,
+    RTVIDescribeActionsData,
+    RTVIDescribeConfig,
+    RTVIDescribeConfigData,
+    RTVIService,
+    RTVIServiceConfig,
+    RTVIServiceOption,
+    RTVIServiceOptionConfig,
+    RTVIUpdateConfig,
+)
+from pipecat.processors.frameworks.rtvi.observer import (
+    RTVIFunctionCallReportLevel,
+    RTVIObserver,
+    RTVIObserverParams,
+)
+from pipecat.processors.frameworks.rtvi.processor import RTVIProcessor
+
+__all__ = [
+    "ActionResult",
+    "RTVIAction",
+    "RTVIActionArgument",
+    "RTVIActionArgumentData",
+    "RTVIActionFrame",
+    "RTVIActionResponse",
+    "RTVIActionResponseData",
+    "RTVIActionRun",
+    "RTVIActionRunArgument",
+    "RTVIBotReadyDataDeprecated",
+    "RTVIClientMessageFrame",
+    "RTVIConfig",
+    "RTVIConfigResponse",
+    "RTVIDescribeActions",
+    "RTVIDescribeActionsData",
+    "RTVIDescribeConfig",
+    "RTVIDescribeConfigData",
+    "RTVIFunctionCallReportLevel",
+    "RTVIObserver",
+    "RTVIObserverParams",
+    "RTVIProcessor",
+    "RTVIServerMessageFrame",
+    "RTVIServerResponseFrame",
+    "RTVIService",
+    "RTVIServiceConfig",
+    "RTVIServiceOption",
+    "RTVIServiceOptionConfig",
+    "RTVIUpdateConfig",
+]
